+++ packages/ssml-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @ssml-builder-js/ssml-core
 
+## 2.18.0
+
 ## 2.17.0
 
 ### Minor Changes